% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-infomat.R
\name{ast-infomat}
\alias{ast-infomat}
\alias{astInfoMat}
\alias{astCov}
\alias{astCor}
\title{Information Matrix Function of Asymmetric Student-t distribution}
\usage{
astInfoMat(pars, data = c(), method = c("expected", "observed"))

astCov(pars)

astCor(pars)
}
\arguments{
\item{pars}{a vector of parameter values for an AST distribution}

\item{data}{a vector of numeric data used to calculate observed information matrix}

\item{method}{one of "expected" and "observed", calculating the expected / observed information matrix}
}
\description{
Information matrix, asymptotic covariance and correlation matrix functions of Asymmetric Student-t distribution
}
\details{
The expected information matrix is calculated by the expectation of the outer product of score functions,
analytical formulas are provided in \emph{Zhu and Galbraith(2010)}.
The observed information matrix is calculated by the expectation of negative Hessian Matrix of the log-likelihood function.
}
\examples{
pars <- c(0.12, 0.6, 0.6, 3, 5)
data <- rast(1000, pars = pars)
round(astInfoMat(pars, data, "observed"), 4)
round(astInfoMat(pars, "expected"), 4)
round(astCov(pars), 4)
round(astCor(pars), 4)
}
\references{
Zhu, D., & Galbraith, J. W. (2010). A generalized asymmetric Student-t distribution with application to financial econometrics. Journal of Econometrics, 157(2), 297-305.\url{https://www.sciencedirect.com/science/article/pii/S0304407610000266}
\url{https://econpapers.repec.org/paper/circirwor/2009s-13.htm}
}
