% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-moments.R
\name{ast-moment}
\alias{ast-moment}
\alias{astMean}
\alias{astVar}
\alias{astSkew}
\alias{astKurt}
\alias{astMoment}
\alias{astMoments}
\alias{astRawMoment}
\alias{astCentralMoment}
\title{Moment Functions of Asymmetric Student-t distribution}
\usage{
astMean(
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu1 = Inf,
  nu2 = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

astMoment(
  moment = c("mean", "sd", "var", "skew", "kurt"),
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu1 = Inf,
  nu2 = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

astMoments(
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu1 = Inf,
  nu2 = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

astRawMoment(
  n,
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu1 = Inf,
  nu2 = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

astCentralMoment(
  n,
  mu,
  s,
  alpha,
  nu1,
  nu2,
  pars = NULL,
  method = c("analytical", "numerical")
)
}
\arguments{
\item{mu}{location parameter}

\item{s}{scale parameter, \eqn{s > 0}}

\item{alpha}{skewness parameter, \eqn{0 < alpha < 1}}

\item{nu1}{degrees of freedom / tail parameter for the left tail, \eqn{ nu1 > 0}}

\item{nu2}{degrees of freedom / tail parameter for the right tail, \eqn{ nu2 > 0}}

\item{pars}{a vector that contains mu, s, alpha, nu1, nu2, if pars is specified, mu, s, alpha, nu1, nu2 should not be specified}

\item{method}{method used to calculate the moment(s), one of 'analytical' and 'numerical'}

\item{moment}{the moment to be calculated, one of 'mean', 'sd', 'skew', 'kurt'}

\item{type}{type of kurtosis calculated, one of 'excess' and 'regular'}

\item{n}{order of (raw/central) moment to be calculated}
}
\description{
The mean, standard deviation, skewness, kurtosis functions, as well as the raw and central moments of AST distribution
}
\details{
Function \code{astMoment} calculates one of mean, standard deviation, skewness and kurtosis of the distribution,
while \code{astMoment} calculates all 4 of them. \cr
Function \code{astRawMoment} returns \eqn{E[X^n]},
while function \code{astCentralMoment} returns \eqn{E[(X-\mu)^n]}

The moments for AST follow the general rule of student t distribution,
\itemize{
    \item mean is only defined for nu > 1,
    \item variance/standard deviation is finite when nu > 2, infinite when 1 < nu < 2, otherwise undefined,
    \item skewness is defined when nu > 3,
    \item kurtosis is finite when nu > 4, infinite when 2 < nu <= 4, otherwise undefined.
}
}
\examples{
# The parameter values are specially set for a volatile portfolio.
pars <- c(0.12, 0.6, 0.6, 6, 5)
astMoment("sd", pars = pars, method = "numerical")
astMoments(pars = pars)
}
\references{
Zhu, D., & Galbraith, J. W. (2010). A generalized asymmetric Student-t distribution with application to financial econometrics. Journal of Econometrics, 157(2), 297-305.\url{https://www.sciencedirect.com/science/article/pii/S0304407610000266}
\url{https://econpapers.repec.org/paper/circirwor/2009s-13.htm}
}
