% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_functions.R
\name{ast}
\alias{ast}
\alias{dast}
\alias{past}
\alias{qast}
\alias{rast}
\title{Asymmetric Student t-distribution Functions}
\usage{
dast(x, mu, sigma, alpha, nu1, nu2)

past(x, location, scale, skewness, dof1, dof2)
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST
}
\examples{
dast()
past()
qast()
rast()
}
