% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astdist.R
\name{astDist}
\alias{astDist}
\alias{dast.numeric}
\alias{dast}
\alias{past.numeric}
\alias{qast.numeric}
\alias{rast.numeric}
\title{Asymmetric Student t-distribution}
\usage{
\method{dast}{numeric}(x, mu, sigma, alpha, nu1, nu2)

\method{past}{numeric}(q, mu, sigma, alpha, nu1, nu2)

\method{qast}{numeric}(p, mu, sigma, alpha, nu1, nu2)

\method{rast}{numeric}(n, mu, sigma, alpha, nu1, nu2)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{alpha}{skewness parameter}

\item{nu1}{degrees of freedom / tail parameter for the left tail}

\item{nu2}{degrees of freedom / tail parameter for the right tail}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST distributions
}
\examples{
# The parameter values are specially set for a volatile portfolio.
d <- dast(0, 0.12, 0.6, 0.3, 3, 5)
p <- past(1.5, 0.12, 0.6, 0.3, 3, 5)
q <- qast(0.8, 0.12, 0.6, 0.3, 3, 5)
x <- rast(1000, 0.12, 0.6, 0.3, 3, 5)
}
