% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astFit-methods.R
\name{astFit-methods}
\alias{astFit-methods}
\alias{summary.astFit}
\alias{moments.astFit}
\alias{print.astFit}
\alias{fitted.astFit}
\alias{se.astFit}
\alias{objective.astFit}
\alias{plot.astFit}
\title{Methods for astFit class}
\usage{
\method{summary}{astFit}(fit)

\method{moments}{astFit}(fit, method = c("analytical", "numerical"))

\method{print}{astFit}(fit)

\method{fitted}{astFit}(fit)

\method{se}{astFit}(fit)

\method{objective}{astFit}(fit)

plot.astFit(fit, type = NULL, dist = "ast", ...)
}
\arguments{
\item{fit}{A AST fit object of class \code{\link{astFit}}}

\item{method}{one of "numerical" and "analytical", calculating the moments using numerical integration / analytical formula}

\item{type}{one of "density" and "QQplot"}
}
\description{
Methods for astFit S3 class
}
\details{
should also add the empirical moments
}
\examples{
pars <- c(0.12, 0.6, 0.6, 6, 5)
data <- rast(1000, pars = pars)
solver_control <- list(eval.max = 10^3, iter.max = 10^3)
fit <- astFit(data, solver = 'nlminb', solver_control = solver_control)
summary(fit)
moments(fit)
fitted(fit)
se(fit)
objective(fit)
plot(fit)
}
