% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astfit-method.R
\name{astfit}
\alias{astfit}
\title{AST fit
the object currently have a S3 structure which will be shifted into S4 in future development}
\usage{
astfit(spec, data, solver, solver.control)
}
\arguments{
\item{spec}{An AST specification object of class \code{\link{astspec}}}

\item{data}{A univariate data object to fit on}

\item{solver}{Optimization algorithm used, one of ...}

\item{solver.control}{Control arguments passed to the optimization algorithm ...}

\item{fit.control}{Control arguments passed to the fitting routine...}
}
\description{
Method for fitting a AST distribution
}
\examples{
spec = astspec(c(0, 1, 0.5, 1, 1))
data = rast(1000, 1.5, 1.2, 0.8, 3, 4)
fit = astfit(spec, data, "nloptr", list("algorithm" = "NLOPT_LN_COBYLA", "maxeval" = 1.0e5, "xtol_rel" = 1.0e-8))
}
