% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-moments.R
\name{astMoments}
\alias{astMoments}
\alias{moment_ast}
\title{Asymmetric Student t-distribution}
\usage{
moment_ast(n, mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf,
  nu2 = Inf, pars = NULL, method = c("analytical", "numerical"))
}
\arguments{
\item{n}{order of moment to be calculated}

\item{mu}{location parameter, the mode, not necessarily the mean}

\item{sigma}{scale parameter, not necessarily the standard deviation, greater than 0}

\item{alpha}{skewness parameter, ranges from 0 to 1, when < 0.5, skew to the right, when > 0.5, skew to the left}

\item{nu1}{degrees of freedom / tail parameter for the left tail, greater than 0}

\item{nu2}{degrees of freedom / tail parameter for the right tail, greater than 0}

\item{pars}{a vector that contains c(mu, sigma, alpha, nu1, nu2), one and only one of {mu, sigma, alpha, nu1, nu2} or pars should be specified}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST distributions
}
\details{
the moments for AST follow the general rule of student t distribution, mean is only defined for nu > 1, variance is finite when nu > 2, infinite when 1 < nu < 2, otherwise undefined; skewness is defined when nu > 3; kurtosis is finite when nu > 4, infinite when 2 < nu <= 4, otherwise undefined (This need more check, and also the analytical methods should be modified)
}
\examples{
# The parameter values are specially set for a volatile portfolio.
pars <- c(0.12, 0.6, 0.4, 3, 5)
moment_ast(1, pars = pars, method = "numerical")
mean_ast(pars = pars)
var_ast(pars = pars)
skew_ast(pars = pars)
kurt_ast(pars = pars)
moments_ast(pars = pars)
}
