% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astspec-method.R
\name{astspec}
\alias{astspec}
\title{AST Specification
the object currently have a S3 structure which may be shifted into S4 in future development}
\usage{
astspec(data, start_pars = c(0, 1, 0.5, 1, 1), fixed_pars = c())
}
\arguments{
\item{data}{A univariate data object, can be ... for the AST distribution to fit on.}

\item{start_pars}{Numeric vector of starting parameters for the optimization algorithm.}

\item{fixed_pars}{Numeric vector of parameters to be kept fixed during the optimization routine.}
}
\description{
Method for creating an AST distribution object prior to fitting.
}
\examples{
data = rast(1000, 1.5, 2, 0.8, 3, 4)
spec = astspec(data)
}
