% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gat-infomat.R
\name{gat-infomat}
\alias{gat-infomat}
\alias{gatInfoMat}
\alias{gatCov}
\alias{gatCor}
\title{Information Matrix Function of Asymmetric Student-t distribution}
\usage{
gatInfoMat(pars, data)

gatCov(pars, data)

gatCor(pars, data)
}
\arguments{
\item{pars}{a vector of parameter values for an AST distribution}

\item{data}{a vector of numeric data used to calculate observed information matrix}
}
\description{
Information matrix, asymptotic covariance and correlation matrix functions of Asymmetric Student-t distribution
}
\details{
The expected information matrix is calculated by the expectation of the outer product of score functions,
analytical formulas are provided in \emph{Zhu and Galbraith(2010)}.
The observed information matrix is calculated by the expectation of negative Hessian Matrix of the log-likelihood function.
}
\examples{
pars <- c(0.12, 0.6, 1.5, 1.2, 2, 5)
data <- rgat(1000, pars = pars)

round(gatInfoMat(pars, data), 4)

round(gatCov(pars, data), 4)
round(gatCor(pars, data), 4)

}
\references{
Baker, R. D. (2016). A new asymmetric generalisation of the t-distribution. arXiv preprint arXiv:1606.05203.
\doi{https://doi.org/10.48550/arXiv.1606.05203}
}
