% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gat-methods.R
\name{gat-methods}
\alias{gat-methods}
\alias{summary.gat}
\alias{moments.gat}
\alias{plots.gat}
\alias{objective.gat}
\alias{print.gat}
\alias{plot.gat}
\title{Methods for gat class}
\usage{
\method{summary}{gat}(object, ...)

\method{moments}{gat}(x, method = c("analytical", "numerical"), ...)

\method{print}{gat}(x, ...)

\method{plot}{gat}(x, type = NULL, dist = "gat", envelope = 0.95, ...)
}
\arguments{
\item{object}{A GAT fit object of class \code{\link{gat}}}

\item{...}{additional arguments for the \code{hist} or \code{plot} function from \code{graphics}}

\item{x}{A GAT fit object of class \code{\link{gat}}}

\item{method}{one of "numerical" and "analytical", calculating the moments using numerical integration / analytical formula}

\item{type}{one of "density" or "qqplot"}

\item{dist}{one of "norm" or "ast", the theoretical distribution used in QQplots}

\item{envelope}{the confidence level used to construct the envelope}
}
\description{
Methods for gat S3 class
}
\details{
should also add the empirical moments
}
\examples{
pars <- c(0.12, 0.6, 1.5, 1.2, 2, 5)
data <- rgat(1000, pars = pars)

fit <- gatMLE(data)

summary(fit)
moments(fit)
plot(fit, 1)

}
