% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gat-moments.R
\name{gat-moment}
\alias{gat-moment}
\alias{gatMean}
\alias{gatVar}
\alias{gatSD}
\alias{gatSkew}
\alias{gatKurt}
\alias{gatMoment}
\alias{gatMoments}
\alias{gatRawMoment}
\alias{gatCentralMoment}
\title{Moment Functions of Asymmetric Student-t distribution}
\usage{
gatMean(
  mu = 0,
  phi = 1,
  alpha = 0.5,
  r = 2,
  c = 2,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

gatMoment(
  moment = c("mean", "sd", "var", "skew", "kurt"),
  mu = 0,
  phi = 1,
  alpha = 0.5,
  r = 2,
  c = 2,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

gatMoments(
  mu = 0,
  phi = 1,
  alpha = 0.5,
  r = 2,
  c = 2,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

gatRawMoment(
  n,
  mu = 0,
  phi = 1,
  alpha = 0.5,
  r = 2,
  c = 2,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

gatCentralMoment(
  n,
  mu = 0,
  phi = 1,
  alpha = 0.5,
  r = 2,
  c = 2,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)
}
\arguments{
\item{mu}{location parameter}

\item{phi}{scale parameter, \eqn{phi > 0}}

\item{alpha}{skewness parameter, \eqn{0 < alpha < 1}}

\item{r}{tail power asymmetry parameter \eqn{r > 0}}

\item{c}{scale asymmetry parameter \eqn{r > 0}}

\item{nu}{degrees of freedom / tail parameter}

\item{pars}{a vector that contains mu, phi, alpha, r, c, nu, if pars is specified, mu, phi, alpha, r, c, nu should not be specified}

\item{method}{method used to calculate the moment(s), one of 'analytical' and 'numerical'}

\item{moment}{the moment to be calculated, one of 'mean', 'sd', 'skew', 'kurt'}

\item{type}{type of kurtosis calculated, one of 'excess' and 'regular'}

\item{n}{order of (raw/central) moment to be calculated}
}
\description{
The mean, standard deviation, skewness, kurtosis functions, as well as the raw and central moments of GAT distribution
}
\details{
Function \code{gatMoment} calculates one of mean, standard deviation, skewness and kurtosis of the distribution,
while \code{gatMoment} calculates all 4 of them. \cr
Function \code{gatRawMoment} returns \eqn{E[X^n]},
while function \code{gatCentralMoment} returns \eqn{E[(X-\mu)^n]}

The moments for GAT follow the general rule of student t distribution,
\itemize{
    \item mean is only defined for nu > 1,
    \item variance/standard deviation is finite when nu > 2, infinite when 1 < nu < 2, otherwise undefined,
    \item skewness is defined when nu > 3,
    \item kurtosis is finite when nu > 4, infinite when 2 < nu <= 4, otherwise undefined.
}
}
\examples{
# The parameter values are specially set for a volatile portfolio.
pars <- c(0.12, 0.6, 1.5, 1.2, 2, 5)

gatMoment("sd", pars = pars, method = "numerical")
gatMoments(pars = pars)
}
\references{
Baker, R. D. (2016). A new asymmetric generalisation of the t-distribution. arXiv preprint arXiv:1606.05203.
\url{https://doi.org/10.48550/arXiv.1606.05203}
}
