% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gat-dist.R
\name{GAT}
\alias{GAT}
\alias{dgat}
\alias{gat}
\alias{pgat}
\alias{qgat}
\alias{rgat}
\title{Generalized Asymmetric t-distribution}
\usage{
dgat(x, mu, scale, alpha, r, c, nu, pars = NULL)

pgat(q, mu, scale, alpha, r, c, nu, pars = NULL)

qgat(p, mu, scale, alpha, r, c, nu, pars = NULL)

rgat(n, mu, scale, alpha, r, c, nu, pars = NULL)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{scale}{scale parameter, \eqn{scale > 0}}

\item{alpha}{how early tail behavior is apparent, \eqn{alpha > 0}}

\item{r}{tail power asymmetry, \eqn{r > 0}}

\item{c}{scale asymmetry, \eqn{c > 0}}

\item{nu}{degrees of freedom / tail parameter, \eqn{nu > 0}}

\item{pars}{a vector that contains mu, scale, alpha, r, c, nu, if pars is specified, mu, scale, alpha, r, c, nu should not be specified}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\value{
\code{dgat} gives the density, \code{pgat} gives the distribution function, \code{qgat} gives the quantile function, and \code{rgat} generates random samples for GATdistribution.
}
\description{
Probablity density function(PDF), Cumulative distribution function(CDF), Quantile function and Random generation of the GAT distribution
}
\examples{
dgat(0, 0.12, 0.6, 1.5, 1.2, 2, 5)
pgat(0.12, 0.12, 0.6, 1.5, 1.2, 2, 5)
qgat(0.4, 0.12, 0.6, 1.5, 1.2, 2, 5)
data = rgat(1000, 0.12, 0.6, 1.5, 1.2, 2, 5)
hist(data, breaks = 50, probability = TRUE)

# using the 'pars' argument
pars <- c(0.12, 0.6, 1.5, 1.2, 2, 5)
x <- seq(-3, 3, 0.01)
y <- dgat(x, pars = pars)
lines(x, y, col = 4)

}
\references{
Baker, R. D. (2016). A new asymmetric generalisation of the t-distribution. arXiv preprint arXiv:1606.05203.
\doi{https://doi.org/10.48550/arXiv.1606.05203}
}
