% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatdist.R
\name{gatDist}
\alias{gatDist}
\alias{dgat}
\alias{pgat}
\alias{rgat}
\title{Generalized Asymmetric t-distribution}
\usage{
dgat(x, mu, sigma, dof, skewness, r, c)

pgat(x, mu, sigma, dof, skewness, r, c)

rgat(n, mu, sigma, dof, skewness, r, c)
}
\arguments{
\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{r}{tail power asymmetry}

\item{c}{scale asymmetry}

\item{n}{number of observations for random generation}

\item{xvector}{of quantiles}

\item{alpha}{how early tail behavior is apparent}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of GAT distributions
}
