% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_matrix.R
\name{infoMat_ast}
\alias{infoMat_ast}
\title{Asymmetric Student t-distribution}
\usage{
infoMat_ast(fit, method = c("expected", "observed"))
}
\arguments{
\item{x, q}{vector of quantiles}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{alpha}{skewness parameter}

\item{nu1}{degrees of freedom / tail parameter 1}

\item{nu2}{degrees of freedom / tail parameter 2}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST distributions
}
