% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{moments}
\alias{moments}
\title{Calculate all moments of a distribution}
\usage{
moments(x, ...)
}
\arguments{
\item{x}{A AST/GAT fit object of class \code{\link{ast}} / \code{\link{gat}}.}

\item{...}{Additional arguments passed into the calculating functions of moments.}
}
\description{
The generic S3 methods for calculating all moments of a fitted distribution.
}
