% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot-methods}
\alias{plot-methods}
\alias{plot_density}
\alias{qqast}
\title{AST plots}
\usage{
plot_density(fit)

qqast(fit, method = "normal")
}
\arguments{
\item{fit}{A AST fit object of class \code{\link{astfit}}}
}
\description{
Plot methods for the AST fit class
}
