% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewtDist-package.R
\docType{package}
\name{skewtDist-package}
\alias{skewtDist}
\alias{skewtDist-package}
\title{skewtDist: Asymmetric Student t-distributions}
\description{
Methods and Applications for Asymmetric Student T-distributions (AST) and Generalized Asymmetric T-distributions (GAT).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/skewtDist}
  \item Report bugs at \url{https://github.com/dan9401/skewtDist/issues}
}

}
\author{
\strong{Maintainer}: Daniel Xia \email{dxia94@gmail.com}

}
\keyword{internal}
