% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst-functions.R
\name{sst-functions}
\alias{sst-functions}
\alias{dsst}
\alias{sst}
\alias{psst}
\alias{qsst}
\alias{rsst}
\alias{sstMoment}
\alias{sstMoments}
\alias{sstRawMoments}
\alias{sstCentralMoments}
\alias{sstMean}
\alias{sstVar}
\alias{sstSD}
\alias{sstSkew}
\alias{sstKurt}
\alias{sstInfoMat}
\alias{sstFit}
\alias{sstRawMoment}
\alias{sstCentralMoment}
\alias{sstMLE}
\title{Functions for Symmetric Student-t distribution}
\usage{
dsst(x, mu = 0, s = 1, alpha = 0.5, nu = Inf, pars = NULL)

psst(q, mu = 0, s = 1, alpha = 0.5, nu = Inf, pars = NULL)

qsst(p, mu = 0, s = 1, alpha = 0.5, nu = Inf, pars = NULL)

rsst(n, mu = 0, s = 1, alpha = 0.5, nu = Inf, pars = NULL)

sstMoments(
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

sstRawMoment(
  n,
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

sstCentralMoment(
  n,
  mu = 0,
  s = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

sstInfoMat(pars)

sstMLE(
  data,
  start_pars = c(),
  fixed_pars = c(),
  solver = c("nlminb", "nloptr", "Rsolnp"),
  solver_control = list()
)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{s}{scale parameter, \eqn{s > 0}}

\item{alpha}{skewness parameter, \eqn{0 < alpha < 1}}

\item{nu}{degrees of freedom / tail parameter for the both tails, \eqn{ nu > 0}}

\item{pars}{a vector that contains mu, s, alpha, nu1, nu2, if pars is specified, mu, s, alpha, nu1, nu2 should not be specified}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}

\item{method}{method used to calculate the moment(s), one of 'analytical' and 'numerical'}

\item{type}{type of kurtosis calculated, one of 'excess' and 'regular'}

\item{data}{a univariate data object to be fitted}

\item{start_pars}{a named numeric vector of starting parameters for the optimization algorithm, not all parameters are needed}

\item{fixed_pars}{a named numeric vector of parameters to be kept fixed during the optimization routine, not all parameters are needed}

\item{solver}{solver used for MLE, one of 'nlminb', 'nloptr', 'Rsolnp', default is 'nlminb'}

\item{solver_control}{list of control arguments passed to the solver}
}
\description{
Functions for Symmetric Student-t distribution
}
\details{
The SST functions are wrappers of their AST equivalent.

For most of the functions, they are implemented by simply setting nu1, nu2(AST) = nu(SST).

For \code{sstFit}, it is implemented by setting symmetric = TRUE.
}
