% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{weeklyRtnData}
\alias{weeklyRtnData}
\alias{retLW,}
\alias{retSW}
\alias{retLW}
\title{Weekly return data for Large and Small Caps stocks from 1997-01-01 to 2010-12-31.}
\format{
\describe{
  \item{retLW}{A xts object of 730 rows and 20 columns, each column represents a different large-cap stock.}
  \item{retSW}{A xts object of 730 rows and 20 columns, each column represents a different small-cap stock.}
}

An object of class \code{xts} (inherits from \code{zoo}) with 730 rows and 20 columns.

An object of class \code{xts} (inherits from \code{zoo}) with 730 rows and 20 columns.
}
\source{
\url{https://finance.yahoo.com/lookup/}
}
\usage{
data(retLW)
       data(retSW)

retLW

retSW
}
\description{
The skewtDist package provides tools for modeling non-normally distributed data using Asymmetric Student-t (AST) and Generalized Asymmetric t-distributions (GAT), which are families of fat-tailed distributions that exhibit different tail behaviors on the left and right sides.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/skewtDist}
  \item Report bugs at \url{https://github.com/dan9401/skewtDist/issues}
}

}
\author{
\strong{Maintainer}: Daniel Xia \email{dxia94@gmail.com}

}
\keyword{datasets}
