% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-methods.R
\name{ast-methods}
\alias{ast-methods}
\alias{summary.ast}
\alias{moments.ast}
\alias{print.ast}
\alias{plot.ast}
\title{Methods for ast class}
\usage{
\method{summary}{ast}(fit)

\method{moments}{ast}(fit, method = c("analytical", "numerical"))

\method{print}{ast}(fit)

plot.ast(fit, type = NULL, dist = "ast", envelope = 0.95, ...)
}
\arguments{
\item{fit}{A AST fit object of class \code{\link{ast}}}

\item{method}{one of "numerical" and "analytical", calculating the moments using numerical integration / analytical formula}

\item{type}{one of "density" and "QQplot"}
}
\description{
Methods for ast S3 class
}
\details{
should also add the empirical moments
}
\examples{
pars <- c(0.12, 0.6, 0.6, 6, 5)
data <- rast(1000, pars = pars)
solver_control <- list(eval.max = 10^3, iter.max = 10^3)
fit <- astMLE(data, solver = 'nlminb', solver_control = solver_control)
summary(fit)
moments(fit)
plot(fit)
}
