% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-dist.R
\name{AST}
\alias{AST}
\alias{dast}
\alias{ast}
\alias{past}
\alias{qast}
\alias{rast}
\title{Asymmetric Student-t Distribution}
\usage{
dast(x, mu = 0, s = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf, pars = NULL)

past(q, mu = 0, s = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf, pars = NULL)

qast(p, mu = 0, s = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf, pars = NULL)

rast(n, mu = 0, s = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf, pars = NULL)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{s}{scale parameter, \eqn{s > 0}}

\item{alpha}{skewness parameter, \eqn{0 < alpha < 1}}

\item{nu1}{degrees of freedom / tail parameter for the left tail, \eqn{ nu1 > 0}}

\item{nu2}{degrees of freedom / tail parameter for the right tail, \eqn{ nu2 > 0}}

\item{pars}{a vector that contains mu, s, alpha, nu1, nu2, if pars is specified, mu, s, alpha, nu1, nu2 should not be specified}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\value{
\code{dast} gives the density, \code{past} gives the distribution function, \code{qast} gives the quantile function, and \code{rast} generates random samples for AST distribution.
}
\description{
Probablity density function(PDF), Cumulative distribution function(CDF), Quantile function and Random generation of the AST distribution
}
\details{
The 'asymmetric' in AST distribution, not only suggests skewness in the distribution,
but also the asymmetry in the two tail powers of the distribution.
\itemize{
    \item Location parameter \code{mu} is the mode, but not necessarily the mean of the distribution.
    \item Scale parameter \code{s} is not necessarily the standard deviation.
    \item The distribution skews to the right when the skewness parameter \eqn{alpha < 0.5},
    skews to the left when \eqn{alpha > 0.5}.
    \item The location paramter \code{mu} always locates at the \eqn{\alpha}-th percentile of the distribution.
    The two degrees of freedom / tail parameters each controls one tail of the distribution,
    separated at the location paramter \code{mu}.
    The left tail parameter \code{nu1} only affects the left half(0th to \eqn{\alpha}-th percentile) of the distribution,
    while the right tail paramter \code{nu2} only affects the right half(\eqn{\alpha}-th to 100-th percentile) of the distribution.
}
}
\examples{
d <- dast(0.12, 0.12, 0.6, 0.6, 3, 5)
p <- past(0.12, 0.12, 0.6, 0.6, 3, 5)
q <- qast(0.4, 0.12, 0.6, 0.6, 3, 5)
data <- rast(1000, 0.12, 0.6, 0.6, 3, 5)
hist(data, breaks = 50, probability = TRUE)

pars <- c(0.12, 0.6, 0.6, 3, 5)
x <- seq(-3, 3, 0.01)
y <- dast(x, pars = pars)
lines(x, y, col = 4)

}
\references{
Zhu, D., & Galbraith, J. W. (2010). A generalized asymmetric Student-t distribution with application to financial econometrics. Journal of Econometrics, 157(2), 297-305.\url{https://www.sciencedirect.com/science/article/pii/S0304407610000266}
\url{https://econpapers.repec.org/paper/circirwor/2009s-13.htm}
}
