% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astdist.R
\name{astDist}
\alias{astDist}
\alias{dast}
\alias{past}
\alias{qast}
\alias{rast}
\title{Asymmetric Student t-distribution}
\usage{
dast(x, mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL)

past(q, mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL)

qast(p, mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL)

rast(n, mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter, the mode, not necessarily the mean}

\item{sigma}{scale parameter, not necessarily the standard deviation, greater than 0}

\item{alpha}{skewness parameter, ranges from 0 to 1, when < 0.5, skew to the right, when > 0.5, skew to the left}

\item{nu1}{degrees of freedom / tail parameter for the left tail, greater than 0}

\item{nu2}{degrees of freedom / tail parameter for the right tail, greater than 0}

\item{pars}{a vector that contains c(mu, sigma, alpha, nu1, nu2), one and only one of {mu, sigma, alpha, nu1, nu2} or pars should be specified}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST distributions
}
\examples{
# The parameter values are specially set for a volatile portfolio.
# density at the mu is always 1 / sigma
d <- dast(0.12, 0.12, 0.6, 0.6, 3, 5)
# cumulative distribution at mu is alpha
p <- past(0.12, 0.12, 0.6, 0.6, 3, 5)
# quantile at alpha is mu
q <- qast(0.4, 0.12, 0.6, 0.6, 3, 5)
data <- rast(1000, 0.12, 0.6, 0.6, 3, 5)
hist(data, breaks = 50, probability = TRUE)

# now trying to use the pars argument
pars <- c(0.12, 0.6, 0.6, 3, 5)
x <- seq(-3, 3, 0.01)
y <- dast(x, pars = pars)
lines(x, y, col = 4)

pars1 <- c(0, 2, 0.3, 5, 5)
pars2 <- c(0, 2, 0.5, 5, 5)
pars3 <- c(0, 2, 0.7, 5, 5)
y1 <- dast(x1, pars = pars1)
y2 <- dast(x1, pars = pars2)
y3 <- dast(x1, pars = pars3)
plot(x, y1, type = "l", main = expression(alpha), xlab = "x", ylab = "density")
lines(x, y2, lty = 2)
lines(x, y3, lty = 3)
abline(v = 0, col = 4, lty = 2)
legend(x = "topleft", legend = c("alpha = 0.3", "alpha = 0.5", "alpha = 0.7"), lty = 1:3)
}
