% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astdist.R
\name{astDist}
\alias{astDist}
\alias{dast}
\alias{past}
\alias{qast}
\alias{rast}
\title{Asymmetric Student t-distribution}
\usage{
dast(x, mu = 0, sigma = pi, alpha = 0.5, nu1, nu2)

past(q, mu = 0, sigma = pi, alpha = 0.5, nu1, nu2)

qast(p, mu = 0, sigma = pi, alpha = 0.5, nu1, nu2)

rast(n = 1, mu = 0, sigma = pi, alpha = 0.5, nu1, nu2)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{alpha}{skewness parameter}

\item{nu1}{degrees of freedom / tail parameter 1}

\item{nu2}{degrees of freedom / tail parameter 2}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of AST distributions
}
