% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astfit-methods.R
\name{astfit-methods}
\alias{astfit-methods}
\alias{summary.astfit}
\alias{moments.astfit}
\alias{plot.astfit}
\title{Methods for astfit class}
\usage{
\method{summary}{astfit}(fit)

\method{moments}{astfit}(fit, method = c("analytical", "numerical"))

plot.astfit(fit, selection = NULL, ...)
}
\arguments{
\item{fit}{A AST fit object of class \code{\link{astfit}}}

\item{method}{one of "numerical" and "analytical", calculating the moments using numerical integration / analytical formula}

\item{selection}{one of 1 & 2, for plotting the density or QQPlot}
}
\description{
Methods for astfit class
}
\details{
should also add the empirical moments
}
\examples{
pars <- c(0.12, 0.6, 0.4, 3, 5)
data <- rast(1000, pars = pars)
solver_control <- list(eval.max = 10^3, iter.max = 10^3)
fit <- astfit(data, solver = 'nlminb', solver_control = solver_control)
summary(fit)
moments(fit)
plot(fit)
}
