% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astfit-method.R
\name{astfit}
\alias{astfit}
\title{AST fit
the object currently have a S3 structure which will be shifted into S4 in future development}
\usage{
astfit(spec, data, solver, solver.control, fit.control)
}
\arguments{
\item{spec}{An AST specification object of class \code{\link{astspec}}}

\item{data}{A univariate data object to fit on}

\item{solver}{Optimization algorithm used, one of ...}

\item{solver.control}{Control arguments passed to the optimization algorithm ...}

\item{fit.control}{Control arguments passed to the fitting routine...}
}
\description{
Method for fitting a AST distribution
}
