% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astfit-method.R
\name{astfit}
\alias{astfit}
\alias{summary.astfit}
\title{AST fit
the object currently have a S3 structure which may be shifted into S4 in future development}
\usage{
astfit(spec, solver, solver_control)

\method{summary}{astfit}(fit)
}
\arguments{
\item{spec}{An AST specification object of class \code{\link{astspec}}}

\item{solver}{Optimizer used for fitting, one of 'nloptr', ...}

\item{solver_control}{Control arguments list passed to the optimizer.}
}
\description{
Method for fitting a AST distribution
}
\examples{
data <- rast(1000, 1.5, 2, 0.8, 3, 4)
spec <- astspec(data)
solver_control <- list('algorithm' = 'NLOPT_LN_COBYLA', 'maxeval' = 1.0e5, 'xtol_rel' = 1.0e-8)
fit <- astfit(spec, 'nloptr', solver_control)
}
