% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astfit.R
\name{astfit}
\alias{astfit}
\title{AST fit}
\usage{
astfit(data, start_pars = c(), fixed_pars = c(), solver = c("nlminb",
  "nloptr", "Rsolnp"), solver_control, symmetric = FALSE)
}
\arguments{
\item{data}{A univariate data object, can be ... for the AST distribution to fit on.}

\item{start_pars}{A named numeric vector of starting parameters for the optimization algorithm.}

\item{fixed_pars}{A named numeric vector of parameters to be kept fixed during the optimization routine.}

\item{solver}{Optimizer used for fitting, one of 'nloptr', 'Rsolnp' ...}

\item{solver_control}{Control arguments list passed to the optimizer.}
}
\description{
Method for fitting a AST distribution
}
\examples{
data <- rast(1000, 0.12, 0.6, 0.3, 3, 5)
# solver_control <- list('algorithm' = 'NLOPT_LN_BOBYQA', 'maxeval' = 1.0e5, 'xtol_rel' = 1.0e-8)
solver_control <- list(eval.max = 10^3, iter.max = 10^3)
# fit <- astfit(data, solver = 'nloptr', solver_control = solver_control)
fit <- astfit(data, solver = 'nlminb', solver_control = solver_control)
}
