% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astspec-method.R
\name{astspec}
\alias{astspec}
\title{AST Specification
the object currently have a S3 structure which will be shifted into S4 in future development}
\usage{
astspec(start.pars = list(), fixed.pars = list())
}
\arguments{
\item{start.pars}{List of starting parameters for the optimization algorithm}

\item{fixed.pars}{List of parameters to be kept fixed during the optimization routine.}
}
\description{
Constructor of a AST distribution object before fitting the parameters
}
