% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astspec-method.R
\name{astspec}
\alias{astspec}
\title{AST Specification
the object currently have a S3 structure which may be shifted into S4 in future development}
\usage{
astspec(data, start_pars = c(mu = 0, sigma = 1, alpha = 0.5, nu1 = 1, nu2
  = 1), fixed_pars = c())
}
\arguments{
\item{data}{A univariate data object, can be ... for the AST distribution to fit on.}

\item{start_pars}{Numeric vector of starting parameters for the optimization algorithm.}

\item{fixed_pars}{Numeric vector of parameters to be kept fixed during the optimization routine.}
}
\description{
Method for creating an AST distribution object prior to fitting.
}
\examples{
data <- rast(1000, 0.12, 0.6, 0.7, 3, 5)
spec <- astspec(data)
}
