% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gat-methods.R
\name{gat-methods}
\alias{gat-methods}
\alias{summary.gat}
\alias{moments.gat}
\alias{print.gat}
\alias{fitted.gat}
\alias{se.gat}
\alias{objective.gat}
\alias{plot.gat}
\title{Methods for gat class}
\usage{
\method{summary}{gat}(fit)

\method{moments}{gat}(fit, method = c("analytical", "numerical"))

\method{print}{gat}(fit)

\method{fitted}{gat}(fit)

\method{se}{gat}(fit)

\method{objective}{gat}(fit)

plot.gat(fit, type = NULL, dist = "gat", envelope, ...)
}
\arguments{
\item{fit}{A GAT fit object of class \code{\link{gat}}}

\item{method}{one of "numerical" and "analytical", calculating the moments using numerical integration / analytical formula}

\item{type}{one of "density" and "QQplot"}
}
\description{
Methods for gat S3 class
}
\details{
should also add the empirical moments
}
\examples{
pars <- c(0.12, 0.6, 0.6, 6, 5)
data <- rgat(1000, pars = pars)
solver_control <- list(eval.max = 10^3, iter.max = 10^3)
fit <- gatMLE(data, solver = 'nlminb', solver_control = solver_control)
summary(fit)
moments(fit)
fitted(fit)
se(fit)
objective(fit)
plot(fit)
}
