% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatMoments.R
\name{gat-moment}
\alias{gat-moment}
\alias{gatMean}
\alias{gatMoments}
\alias{gatRawMoment}
\alias{gatCentralMoment}
\title{Moment Functions of Asymmetric Student-t distribution}
\usage{
gatMean(mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL, method = c("analytical", "numerical"))

gatMoments(mu = 0, sigma = 1, alpha = 0.5, nu1 = Inf, nu2 = Inf,
  pars = NULL, method = c("analytical", "numerical"),
  type = c("excess", "regular"))

gatRawMoment(n, mu = 0, phi = 1, alpha = 0.5, r = 2, c = 2,
  nu = Inf, pars = NULL, method = c("analytical", "numerical"))

gatCentralMoment(n, mu = 0, phi = 1, alpha = 0.5, r = 2, c = 2,
  nu = Inf, pars = NULL, method = c("analytical", "numerical"))
}
\arguments{
\item{mu}{location parameter}

\item{sigma}{scale parameter, \eqn{sigma > 0}}

\item{alpha}{skewness parameter, \eqn{0 < alpha < 1}}

\item{nu1}{degrees of freedom / tail parameter for the left tail, \eqn{ nu1 > 0}}

\item{nu2}{degrees of freedom / tail parameter for the right tail, \eqn{ nu2 > 0}}

\item{pars}{a vector that contains mu, phi, alpha, r, c, nu, if pars is specified, mu, phi, alpha, r, c, nu should not be specified}

\item{method}{method used to calculate the moment(s), one of 'analytical' and 'numerical'}

\item{type}{type of kurtosis calculated, one of 'excess' and 'regular'}

\item{n}{order of (raw/central) moment to be calculated}

\item{moment}{the moment to be calculated, one of 'mean', 'sd', 'skew', 'kurt'}
}
\description{
The mean, standard deviation, skewness, kurtosis functions, as well as the raw and central moments of GAT distribution
}
\details{
Function \code{gatMoment} calculates one of mean, standard deviation, skewness and kurtosis of the distribution,
while \code{gatMoment} calculates all 4 of them. \cr
Function \code{gatRawMoment} returns \eqn{E[X^n]},
while function \code{gatCentralMoment} returns \eqn{E[(X-\mu)^n]}

The moments for GAT follow the general rule of student t distribution,
\itemize{
    \item mean is only defined for nu > 1,
    \item variance/standard deviation is finite when nu > 2, infinite when 1 < nu < 2, otherwise undefined,
    \item skewness is defined when nu > 3,
    \item kurtosis is finite when nu > 4, infinite when 2 < nu <= 4, otherwise undefined.
}
}
\examples{
# The parameter values are specially set for a volatile portfolio.
pars <- c(0.12, 0.6, 0.6, 6, 5)
gatMoment("sd", pars = pars, method = "numerical")
gatMoments(pars = pars)
}
\references{
Zhu, D., & Galbraith, J. W. (2010). A generalized asymmetric Student-t distribution with application to financial econometrics. Journal of Econometrics, 157(2), 297-305.\url{https://www.sciencedirect.com/science/article/pii/S0304407610000266}
\url{https://econpapers.repec.org/paper/circirwor/2009s-13.htm}
}
