% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatdist.R
\name{gatDist}
\alias{gatDist}
\alias{dgat}
\alias{pgat}
\alias{qgat}
\alias{rgat}
\title{Generalized Asymmetric t-distribution}
\usage{
dgat(x, mu, sigma, alpha, r, c, nu)

pgat(q, mu, sigma, alpha, r, c, nu)

qgat(p, mu, sigma, alpha, r, c, nu)

rgat(n, mu, sigma, alpha, r, c, nu)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{alpha}{how early tail behavior is apparent}

\item{r}{tail power asymmetry}

\item{c}{scale asymmetry}

\item{nu}{degrees of freedom / tail parameter}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of GAT distributions
}
