% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatdist.R
\name{gatDist}
\alias{gatDist}
\alias{dgat}
\alias{pgat}
\alias{qgat}
\alias{rgat}
\title{Generalized Asymmetric t-distribution}
\usage{
dgat(x, mu, phi, alpha, r, c, nu)

pgat(x, mu, phi, alpha, r, c, nu)

qgat(p, mu, phi, alpha, r, c, nu)

rgat(n, mu, phi, alpha, r, c, nu)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{mu}{location parameter}

\item{phi}{scale parameter}

\item{alpha}{how early tail behavior is apparent}

\item{r}{tail power asymmetry}

\item{c}{scale asymmetry}

\item{nu}{degrees of freedom / tail parameter}

\item{p}{vector of probablilities}

\item{n}{number of observations for random generation}
}
\description{
Probablity density function (pdf), Cumulative distribution function (cdf), quantile function and random generation of GAT distributions
}
\examples{
# The parameter values are specially set for a volatile portfolio.
# The sample code are not yet decided because of the incompleteness
# of code for GAT distributions.
d <- dgat(0, 0.12, 0.6, 0.3, 3, 5)
p <- pgat(1.5, 0.12, 0.6, 0.3, 3, 5)
q <- qgat(0.8, 0.12, 0.6, 0.3, 3, 5)
x <- rgat(1000, 0.12, 0.6, 0.3, 3, 5)
}
