% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-infomat.R
\name{infoMat_ast}
\alias{infoMat_ast}
\alias{infoMat_sst}
\title{Information Matrix for AST and GAT distributions}
\usage{
infoMat_ast(pars, data = c(), method = c("expected", "observed"))

infoMat_sst(pars)
}
\arguments{
\item{pars}{vector of parameter values for an AST distribution (or an GAT distribution)}

\item{data}{vector of numeric data}

\item{method}{one of "expected" and "observed", calculating the expected / observed information matrix}
}
\description{
Information Matrix function for AST and GAT distributions, function for GAT distributions have not yet been developed
we may want information matrix for symmetric t distributions, or functions for ast & gat objects without data, for users to explore the qualities of distribuitons
we may also want to keep separate files for both distributions, doesn't seem necessary at the time
and infoMat functions on fit and dist objects
}
\examples{
pars <- c(mu = 0.12, sigma = 0.6, alpha = 0.7, nu1 = 3, nu2 = 5)
data <- rast(1000, 0.12, 0.6, 0.3, 3, 5)
infoMat <- infoMat_ast(pars, data, "expected")
}
