% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast-infomat.R
\name{infoMat_ast}
\alias{infoMat_ast}
\alias{infoMat_sst}
\title{Information Matrix for ASTdistributions}
\usage{
infoMat_ast(pars, data = c(), method = c("expected", "observed"))

infoMat_sst(pars)
}
\arguments{
\item{pars}{vector of parameter values for an AST distribution}

\item{data}{vector of numeric data}

\item{method}{one of "expected" and "observed", calculating the expected / observed information matrix}
}
\description{
Information Matrix for AST distributions
}
\details{
the observed method for sst is not yet implemented
}
\examples{
pars <- c(0.12, 0.6, 0.4, 3, 5)
data <- rast(1000, 0.12, 0.6, 0.4, 3, 5)
infoMat_ast(pars, data, "observed")
infoMat_ast(pars, "expected")
pars_s <- c(0.12, 0.6, 0.4, 4)
infoMat_sst(pars)
}
