% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_matrix.R
\name{infoMat_ast}
\alias{infoMat_ast}
\title{Information Matrix for AST distributions}
\usage{
infoMat_ast(pars, data = c(), method = c("expected", "observed"))
}
\arguments{
\item{pars}{vector of parameter values for an AST distribution}

\item{data}{vector of numeric data}

\item{method}{one of "expected" and "observed", calculating the expected or observed information matrix}
}
\description{
Information Matrix function for AST distributions
}
\examples{
pars <- c("mu" = 0.12, "sigma" = 0.6, "alpha" = 0.7, "nu1" = 3, "nu2" = 5)
data <- rast(1000, 0.12, 0.6, 0.7, 3, 5)
infoMat <- infoMat_ast(pars, data, "expected")
}
