% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot-methods}
\alias{plot-methods}
\alias{plot.astfit}
\title{AST plots}
\usage{
\method{plot}{astfit}(fit)
}
\arguments{
\item{fit}{A AST fit object of class \code{\link{astfit}}}
}
\description{
Plot methods for the AST fit class
}
\examples{
data = rast(1000, 1.5, 2, 0.8, 3, 4)
spec <- astspec(data)
solver_control <- list('algorithm' = 'NLOPT_LN_COBYLA', 'maxeval' = 1.0e5, 'xtol_rel' = 1.0e-8)
fit <- astfit(spec, 'nloptr', solver_control)
plot(fit)
}
