% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-package.R
\docType{package}
\name{skewtDist-package}
\alias{skewtDist}
\alias{skewtDist-package}
\title{skewtDist: Asymmetric Student t-distributions}
\description{
Methods and Applications for AST and GAT distributions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/skewtDist}
  \item Report bugs at \url{https://github.com/dan9401/skewtDist/issues}
}

}
\keyword{internal}
