% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewtDist-package.R
\docType{package}
\name{skewtDist-package}
\alias{skewtDist}
\alias{skewtDist-package}
\title{skewtDist: Asymmetric Student t-Distributions}
\description{
The skewtDist package provides tools for modeling non-normally distributed data using Asymmetric Student-t (AST) and Generalized Asymmetric t-distributions (GAT), which are families of fat-tailed distributions that exhibit different tail behaviors on the left and right sides.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/skewtDist}
  \item Report bugs at \url{https://github.com/dan9401/skewtDist/issues}
}

}
\author{
\strong{Maintainer}: Daniel Xia \email{dxia94@gmail.com}

}
\keyword{internal}
