% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewtDist-package.R
\docType{package}
\name{skewtDist-package}
\alias{skewtDist}
\alias{skewtDist-package}
\title{skewtDist: Asymmetric Student t-distributions}
\description{
This package is is a collection of methods and applications for AST distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/st}
  \item Report bugs at \url{https://github.com/dan9401/st/issues}
}

}
\keyword{internal}
