% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst-functions.R
\name{sst-functions}
\alias{sst-functions}
\alias{dsst}
\alias{sst}
\alias{psst}
\alias{qsst}
\alias{rsst}
\alias{sstMoment}
\alias{sstMoments}
\alias{sstRawMoment}
\alias{sstCentralMoment}
\alias{sstMean}
\alias{sstSD}
\alias{sstVar}
\alias{sstSkew}
\alias{sstKurt}
\alias{sstInfoMat}
\alias{sstFit}
\title{Functions for Symmetric Student-t distribution}
\usage{
dsst(x, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

psst(q, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

qsst(p, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

rsst(n, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

sstMoment(moment, mu = 0, sigma = 1, alpha = 0.5, nu = Inf,
  pars = NULL, method = c("analytical", "numerical"),
  type = c("excess", "regular"))

sstMoments(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"), type = c("excess", "regular"))

sstRawMoment(n, mu = 0, sigma = 1, alpha = 0.5, nu = Inf,
  pars = NULL, method = c("analytical", "numerical"))

sstCentralMoment(n, mu = 0, sigma = 1, alpha = 0.5, nu = Inf,
  pars = NULL, method = c("analytical", "numerical"))

sstMean(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"))

sstSD(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"))

sstVar(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"))

sstSkew(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"))

sstKurt(mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL,
  method = c("analytical", "numerical"), type = c("excess", "regular"))

sstInfoMat(pars)

sstFit(data, start_pars = c(), fixed_pars = c(), solver = c("nlminb",
  "nloptr", "Rsolnp"), solver_control = list())
}
\description{
Functions for Symmetric Student-t distribution
}
\details{
The SST functions are wrappers of their AST equivalent.

For most of the functions, they are implemented by simply setting nu1, nu2(AST) = nu(SST).

For \code{sstFit}, it is implemented by setting symmetric = TRUE.
}
