% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst-functions.R
\name{sst-functions}
\alias{sst-functions}
\alias{dsst}
\alias{sst}
\alias{psst}
\alias{qsst}
\alias{rsst}
\alias{sstMoment}
\alias{sstMoments}
\alias{sstRawMoments}
\alias{sstCentralMoments}
\alias{sstMean}
\alias{sstVar}
\alias{sstSD}
\alias{sstSkew}
\alias{sstKurt}
\alias{sstInfoMat}
\alias{sstFit}
\alias{sstRawMoment}
\alias{sstCentralMoment}
\alias{sstMLE}
\title{Functions for Symmetric Student-t distribution}
\usage{
dsst(x, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

psst(q, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

qsst(p, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

rsst(n, mu = 0, sigma = 1, alpha = 0.5, nu = Inf, pars = NULL)

sstMoments(
  mu = 0,
  sigma = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical"),
  type = c("excess", "regular")
)

sstRawMoment(
  n,
  mu = 0,
  sigma = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

sstCentralMoment(
  n,
  mu = 0,
  sigma = 1,
  alpha = 0.5,
  nu = Inf,
  pars = NULL,
  method = c("analytical", "numerical")
)

sstInfoMat(pars)

sstMLE(
  data,
  start_pars = c(),
  fixed_pars = c(),
  solver = c("nlminb", "nloptr", "Rsolnp"),
  solver_control = list()
)
}
\description{
Functions for Symmetric Student-t distribution
}
\details{
The SST functions are wrappers of their AST equivalent.

For most of the functions, they are implemented by simply setting nu1, nu2(AST) = nu(SST).

For \code{sstFit}, it is implemented by setting symmetric = TRUE.
}
