% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st-package.R
\docType{package}
\name{st-package}
\alias{st}
\alias{st-package}
\title{st: Asymmetric Student t-distributions}
\description{
Methods and Applications for AST and GAT distributions
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dan9401/st}
  \item Report bugs at \url{https://github.com/dan9401/st/issues}
}

}
\keyword{internal}
